backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.9.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
pydantic[email]==2.5.0

# Basic dependencies
numpy==1.24.3

# Semantic search dependencies
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# Postgres + pgvector
psycopg2-binary==2.9.9
pgvector==0.2.5
psutil==5.9.8